add support for jsx instead of virtual-dom
