+++ frontend/.babelrc
@@ -1,0 +1,5 @@
+{
+  "plugins": [
+    ["transform-react-jsx", { "pragma": "hJSX" }]
+  ]
+}
