+++ frontend/package.json
@@ -14,11 +14,14 @@
     "rx": "^4.0.6"
   },
   "devDependencies": {
+    "babel-plugin-transform-react-jsx": "^6.6.5",
     "babel-preset-es2015": "^6.1.2",
     "gulp": "^3.9.1",
     "gulp-babel": "^6.1.0",
-    "gulp-serve": "^3.1.1",
+    "gulp-batch": "^1.0.5",
     "gulp-livereload": "^3.8.1",
+    "gulp-plumber": "^1.1.0",
+    "gulp-serve": "^1.2.0",
     "gulp-watch": "^4.3.5",
     "webpack": "^1.12.3"
   }
