+++ frontend/src/index.jsx
@@ -1,0 +1,22 @@
+import Cycle from '@cycle/core';
+import {makeDOMDriver, hJSX} from '@cycle/dom';
+
+function main(drivers) {
+  return {
+    DOM: drivers.DOM.select('input').events('click')
+      .map(ev => ev.target.checked)
+      .startWith(false)
+      .map(toggled =>
+        <div>
+          <label><input type="checkbox" /> Click me</label>
+          <p>{toggled ? 'ON' : 'off'}</p>
+        </div>
+      )
+  };
+}
+
+const drivers = {
+  DOM: makeDOMDriver('#app')
+};
+
+Cycle.run(main, drivers);
